requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0